fix: create separate component for form preview
