+++ src/components/Creator/FormPreview/index.tsx
@@ -1,1 +1,1 @@
-export * from "./FormPreview";
+export * from "./FormPreviewLanding";
